03-02-22-1: Implementation of search operation functionality using filters. Minor UI improvements.

+++ app/src/main/java/com/example/hikingapp/persistence/RouteInfo.kt
@@ -5,8 +5,8 @@ import com.example.hikingapp.domain.DifficultyLevel
 import com.example.hikingapp.domain.RouteType
 
 class RouteInfo(
-    var distance: Int,
-    var timeEstimation: Double,
+    var distance: Double, // in meters
+    var timeEstimation: Double, // in minutes
     var routeType: RouteType,
     var difficultyLevel: DifficultyLevel,
     var rating: Float
